# docs/api/openapi.yaml
info:
  description: 'Human-facing REST API for QuantDinger. Agent integrations use `/api/agent/v1`
    — see `docs/agent/agent-openapi.json`. Conventions: `docs/architecture/API_CONVENTIONS.md`.'
  contact:
    name: QuantDinger
    url: https://github.com/OpenByteInc/quantdinger
  license:
    name: See repository LICENSE
  title: QuantDinger Web API
  version: 1.0.0
components:
  securitySchemes:
    HumanJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'User session JWT from `POST /api/auth/login`. Send as `Authorization:
        Bearer <token>`.'
    AgentToken:
      type: http
      scheme: bearer
      bearerFormat: qd_agent_xxx
      description: Agent Gateway token for `/api/agent/v1/*` only. See docs/agent/AGENT_QUICKSTART.md.
  schemas:
...
    HumanSuccessEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 1 on success
          example: 1
        msg:
          type: string
          example: success
        data:
          nullable: true
      additionalProperties: false
    HumanErrorEnvelope:
      type: object
      properties:
        code:
          type: integer
          description: 0 on error
          example: 0
        msg:
          type: string
        data:
          nullable: true
      additionalProperties: false
...
servers:
- url: http://localhost:5000
  description: Backend direct (python run.py)
- url: http://localhost:8888
  description: Docker Compose (via reverse proxy)
paths:
...
  /api/account/positions:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: L1 account position mirror for Quick Trade / asset views.
      tags:
      - Account
      operationId: getApiAccountPositions
      x-visibility: internal
...
  /api/ibkr/status:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get IBKR connection status.
      tags:
      - IBKR
      operationId: getApiIbkrStatus
      x-visibility: internal
  /api/ibkr/connect:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Connect to TWS / IB Gateway.
      description: 'Request body:

        host (optional, default 127.0.0.1): TWS/Gateway host

        port (optional, default 7497): TWS Live 7496, TWS Paper 7497, Gateway Live
        4001, Gateway Paper 4002

        clientId (optional, default 1): Client ID

        account (optional): Account for multi-account setups

        readonly (optional, default false): Read-only mode'
      tags:
      - IBKR
      operationId: postApiIbkrConnect
      x-visibility: internal
  /api/ibkr/disconnect:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Disconnect from IBKR.
      tags:
      - IBKR
      operationId: postApiIbkrDisconnect
      x-visibility: internal
  /api/ibkr/account:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get IBKR account information.
      tags:
      - IBKR
      operationId: getApiIbkrAccount
      x-visibility: internal
  /api/ibkr/positions:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get IBKR positions.
      tags:
      - IBKR
      operationId: getApiIbkrPositions
      x-visibility: internal
  /api/ibkr/orders:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Get open IBKR orders.
      tags:
      - IBKR
      operationId: getApiIbkrOrders
      x-visibility: internal
  /api/ibkr/order:
    post:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Place an IBKR order.
      description: 'Request body:

        symbol (required): Ticker, e.g. AAPL

        side (required): buy or sell

        quantity (required): Number of shares

        marketType (optional, default USStock): Market type

        orderType (optional, default market): market or limit

        price (required for limit orders): Limit price'
      tags:
      - IBKR
      operationId: postApiIbkrOrder
      x-visibility: internal
  /api/ibkr/order/{order_id}:
    delete:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: Cancel an IBKR order by ID.
      tags:
      - IBKR
      operationId: deleteApiIbkrOrderOrderId
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: integer
      x-visibility: internal
    parameters:
    - in: path
      name: order_id
      required: true
      schema:
        type: integer
        minimum: 0
  /api/ibkr/quote:
    get:
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanErrorEnvelope'
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HumanSuccessEnvelope'
      summary: 'Get real-time IBKR quote (query: symbol, marketType).'
      tags:
      - IBKR
      operationId: getApiIbkrQuote
      x-visibility: internal
...
openapi: 3.0.3